➖ Removes FlutterApi dependency from AdListeners
This change refactors the XandrAdListener classes to remove the direct dependency on `XandrFlutterApi`.

The `flutterApi` instance was previously passed to the `XandrAdListener` and its subclasses (`XandrInterstitialAdListener`, `XandrBannerAdListener`). This dependency is no longer needed as the listeners communicate events via `EventSink`.

Removes the `flutterApi` from `FlutterState` as it's no longer needed.

Also adds a null check to the `invokeOnCompletion` block in `XandrPlugin` to ensure that errors during initialization are properly handled.